# Conflicts:
#	app/src/main/java/com/example/sunandmoon/MainActivity.kt
#	app/src/main/java/com/example/sunandmoon/data/ProductionSelectionUIState.kt
#	app/src/main/java/com/example/sunandmoon/ui/screens/ProductionSelectionScreen.kt

+++ app/src/main/java/com/example/sunandmoon/model/LocationTimeZoneOffsetResultModel/LocationTimeZoneOffsetResult.kt
@@ -7,7 +7,7 @@ data class LocationTimeZoneOffsetResult(
     val latitude: String,
     val longitude: String,
     val timezone_id: String,
-    val offset: Int,
+    val offset: Double,
     val country_code: String,
     val map_url: String
 )
